Correcciones de medición de tiempo"Complejidad"
Se corrigieron errores de medición de tiempo, además se agregaron nuevos archivos para las pruebas de 10mil, 100 mil, 500mil, 1 millon y 10 millones. (A partir de los 10millones se acaba la memoria.)

+++ src/DataSrc/Estacion.java
@@ -35,7 +35,7 @@ public class Estacion implements Comparable<Estacion> {
         this.Nombre = Nombre;
     }
     
-     public Estacion(String Nombre, int Nvagones, int NEntradas, long latitude, long longitude) {
+     public Estacion(String Nombre, int Nvagones, int NEntradas, float latitude, float longitude) {
         this.Nombre = Nombre;
         this.Nvagones = Nvagones;
         this.NEntradas = NEntradas;
